+++ components/card_footer.js
@@ -8,6 +8,7 @@ import EmbeddedBenefitCard from "./embedded_benefit_card";
 import KeyboardReturnIcon from "@material-ui/icons/KeyboardReturn";
 import { cx, css } from "react-emotion";
 import { Grid } from "@material-ui/core";
+import { connect } from "react-redux";
 
 const headerDesc = css`
   flex-grow: 1;
@@ -65,18 +66,65 @@ export class CardFooter extends Component {
   state = {
     open: false
   };
+  benefitTitle = benefit => {
+    return this.props.t("current-language-code") === "en"
+      ? benefit.vacNameEn
+      : benefit.vacNameFr;
+  };
+  childBenefitNames = (benefit, childBenefits, open) => {
+    const length = childBenefits.length;
+    if (open) {
+      return this.props.t("benefits_b.eligible_open_veteran", {
+        x: this.benefitTitle(benefit)
+      });
+    } else {
+      if (length === 1) {
+        return (
+          this.benefitTitle(benefit) +
+          " " +
+          this.props.t("benefits_b.eligible_for_single", {
+            x: this.benefitTitle(childBenefits[0])
+          })
+        );
+      } else {
+        return (
+          this.benefitTitle(benefit) +
+          " " +
+          this.props.t("benefits_b.eligible_for_multi", {
+            x: length
+          })
+        );
+      }
+    }
+  };
+  toggleOpenState = () => {
+    this.setState(previousState => {
+      return { ...previousState, open: !previousState.open };
+    });
+  };
+
   render() {
-    const { t, benefit } = this.props;
+    const { t, benefit, benefits } = this.props;
+
+    let veteranBenefitIds = [];
+    let familyBenefitIds = [];
+
+    this.props.eligibilityPaths.forEach(ep => {
+      if (ep.patronType === "service-person") {
+        veteranBenefitIds = veteranBenefitIds.concat(ep.benefits);
+      }
+      if (ep.patronType === "family") {
+        familyBenefitIds = familyBenefitIds.concat(ep.benefits);
+      }
+    });
     const childBenefits = benefit.childBenefits
-      ? this.props.allBenefits.filter(
-          ab => benefit.childBenefits.indexOf(ab.id) > -1
-        )
+      ? benefits.filter(ab => benefit.childBenefits.indexOf(ab.id) > -1)
       : [];
     const veteranBenefits = childBenefits.filter(
-      ab => this.props.veteranBenefitIds.indexOf(ab.id) > -1
+      ab => veteranBenefitIds.indexOf(ab.id) > -1
     );
     const familyBenefits = childBenefits.filter(
-      ab => this.props.familyBenefitIds.indexOf(ab.id) > -1
+      ab => familyBenefitIds.indexOf(ab.id) > -1
     );
 
     if (childBenefits.length > 0) {
@@ -161,12 +209,19 @@ export class CardFooter extends Component {
     }
   }
 }
-
+const mapStateToProps = reduxState => {
+  return {
+    eligibilityPaths: reduxState.eligibilityPaths,
+    benefits: reduxState.benefits
+  };
+};
 CardFooter.propTypes = {
   showFavourite: PropTypes.bool.isRequired,
-  benefit: PropTypes.array.isRequired,
+  benefit: PropTypes.object.isRequired,
+  benefits: PropTypes.array.isRequired,
   t: PropTypes.func.isRequired,
+  eligibilityPaths: PropTypes.array.isRequired,
   store: PropTypes.object
 };
 
-export default CardFooter;
+export default connect(mapStateToProps)(CardFooter);
